Add initial support for dynamic theming with pywal.
`pywal` provides the tools to theme terminal emulators and other
applications consistently. This commit adds initial support for
dynamically selecting and changing the theme without having to modify
the config file or even restart qtile. Window borders are not yet
supported. The `~/.cache/wal/colors.json` is not watched for changes.
Those points are still to do.